+++ internal/cert/payload.go
@@ -16,11 +16,11 @@ import (
 )
 
 type ConfigPayload struct {
-	CertID                  int                        `json:"cert_id"`
+	CertID                  uint64                     `json:"cert_id"`
 	ServerName              []string                   `json:"server_name"`
 	ChallengeMethod         string                     `json:"challenge_method"`
-	DNSCredentialID         int                        `json:"dns_credential_id"`
-	ACMEUserID              int                        `json:"acme_user_id"`
+	DNSCredentialID         uint64                     `json:"dns_credential_id"`
+	ACMEUserID              uint64                     `json:"acme_user_id"`
 	KeyType                 certcrypto.KeyType         `json:"key_type"`
 	Resource                *model.CertificateResource `json:"resource,omitempty"`
 	MustStaple              bool                       `json:"must_staple"`
