+++ internal/middleware/proxy_ws.go
@@ -16,7 +16,7 @@ func ProxyWs() gin.HandlerFunc {
 			c.Next()
 			return
 		}
-		id := cast.ToInt(nodeID)
+		id := cast.ToUint64(nodeID)
 		if id == 0 {
 			c.Next()
 			return
